# models/User.js
const { DataTypes } = require('sequelize');
const {sq} = require('./db');

const User = sq.define('User', {
  email: {
    type: DataTypes.STRING,
    allowNull: false,
    unique: true,
  },
  password: {
    type: DataTypes.STRING,
    allowNull: false,
  },
});

module.exports = User;   